Bug 818218 - post: Fold TestJSONParsing.java into TestExtendedJSONObject.java.
